# packages/contracts/survey-backend.openapi.yaml
openapi: 3.0.3
info:
  title: Survey Backend API
  version: 1.0.0
  description: API for managing surveys and collecting patient/survey responses.
servers:
  - url: http://localhost:8000/api/v1
paths:
  /surveys/:
    get:
      summary: List surveys
      operationId: listSurveys
      responses:
        '200':
          description: List of surveys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Survey'
    post:
      summary: Create survey
      operationId: createSurvey
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Survey'
      responses:
        '200':
          description: Created survey
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Survey'
  /surveys/{surveyId}:
    get:
      summary: Get survey by id
      operationId: getSurvey
      parameters:
        - in: path
          name: surveyId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Survey found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Survey'
        '404':
          description: Survey not found
  /survey_responses/:
    get:
      summary: List survey responses
      operationId: listSurveyResponses
      responses:
        '200':
          description: List of survey responses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SurveyResponse'
    post:
      summary: Create survey response
      operationId: createSurveyResponse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SurveyResponse'
      responses:
        '201':
          description: Created survey response with optional agent output
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SurveyResponseWithAgent'
  /survey_responses/{responseId}:
    get:
      summary: Get survey response by id
      operationId: getSurveyResponse
      parameters:
        - in: path
          name: responseId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Survey response found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SurveyResponse'
        '404':
          description: Survey response not found
  /survey_responses/{responseId}/send_email:
    post:
      summary: Resend survey response email
      operationId: resendSurveyEmail
      parameters:
        - in: path
          name: responseId
          required: true
          schema:
            type: string
      responses:
        '202':
          description: Email dispatch started
  /patient_responses/:
    post:
      summary: Create patient response
      operationId: createPatientResponse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SurveyResponse'
      responses:
        '201':
          description: Created patient response with optional agent output
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SurveyResponseWithAgent'
  /clinical_writer/process:
    post:
      summary: Forward content to Clinical Writer
      operationId: processClinicalWriter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClinicalWriterRequest'
      responses:
        '200':
          description: Clinical Writer response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentResponse'
  /screeners/register:
    post:
      summary: Register a new screener
      operationId: registerScreener
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScreenerRegister'
      responses:
        '201':
          description: Screener registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScreenerModel'
        '409':
          description: Conflict (e.g., CPF or email already registered)
  /screeners/login:
    post:
      summary: Authenticate a screener and get an access token
      operationId: loginScreener
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScreenerLogin'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '401':
          description: Invalid credentials
  /screeners/recover-password:
    post:
      summary: Request password recovery for a screener
      operationId: recoverScreenerPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScreenerPasswordRecoveryRequest'
      responses:
        '200':
          description: Password recovery initiated (email sent if user exists)
        '500':
          description: Internal server error or email service not configured
  /screeners/me:
    get:
      summary: Get the current screener profile
      operationId: getCurrentScreener
      parameters:
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
          description: Bearer token in the format `Bearer <token>`
      responses:
        '200':
          description: Current screener profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScreenerProfile'
        '401':
          description: Not authenticated or invalid token
        '404':
          description: Screener not found
components:
  schemas:
    Survey:
      type: object
      required:
        - surveyDisplayName
        - surveyName
        - surveyDescription
        - creatorId
        - createdAt
        - modifiedAt
        - instructions
        - questions
        - finalNotes
      properties:
        _id:
          type: string
        surveyDisplayName:
          type: string
        surveyName:
          type: string
        surveyDescription:
          type: string
        creatorId:
          type: string
        createdAt:
          type: string
          format: date-time
        modifiedAt:
          type: string
          format: date-time
        instructions:
          $ref: '#/components/schemas/Instructions'
        questions:
          type: array
          items:
            $ref: '#/components/schemas/Question'
        finalNotes:
          type: string
    Instructions:
      type: object
      required:
        - preamble
        - questionText
        - answers
      properties:
        preamble:
          type: string
        questionText:
          type: string
        answers:
          type: array
          items:
            type: string
    Question:
      type: object
      required:
        - id
        - questionText
        - answers
      properties:
        id:
          type: integer
        questionText:
          type: string
        answers:
          type: array
          items:
            type: string
    Patient:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        birthDate:
          type: string
        age:
          type: integer
        gender:
          type: string
        ethnicity:
          type: string
        educationLevel:
          type: string
        profession:
          type: string
        medication:
          type: array
          items:
            type: string
        diagnoses:
          type: array
          items:
            type: string
        family_history:
          type: string
        social_history:
          type: string
        medical_history:
          type: string
        medication_history:
          type: string
    Answer:
      type: object
      properties:
        id:
          type: integer
        answer:
          type: string
    SurveyResponse:
      type: object
      required:
        - surveyId
        - answers
        - creatorId
        - screenerId
      properties:
        _id:
          type: string
        surveyId:
          type: string
        creatorId:
          type: string
        testDate:
          type: string
          format: date-time
        screenerId:
          type: string
        patient:
          $ref: '#/components/schemas/Patient'
        answers:
          type: array
          items:
            $ref: '#/components/schemas/Answer'
    AgentResponse:
      type: object
      properties:
        ok:
          type: boolean
        input_type:
          type: string
        prompt_version:
          type: string
        model_version:
          type: string
        report:
          type: object
        warnings:
          type: array
          items:
            type: string
        classification:
          type: string
        medicalRecord:
          type: string
        errorMessage:
          type: string
    ClinicalWriterRequest:
      type: object
      required:
        - input_type
        - content
        - locale
        - prompt_key
        - output_format
        - metadata
      properties:
        input_type:
          type: string
          enum:
            - consult
            - survey7
            - full_intake
        content:
          type: string
        locale:
          type: string
        prompt_key:
          type: string
        output_format:
          type: string
          enum:
            - report_json
        metadata:
          type: object
          properties:
            source_app:
              type: string
            request_id:
              type: string
            patient_ref:
              type: string
    SurveyResponseWithAgent:
      allOf:
        - $ref: '#/components/schemas/SurveyResponse'
        - type: object
          properties:
            agentResponse:
              $ref: '#/components/schemas/AgentResponse'
    Address:
      type: object
      required:
        - postalCode
        - street
        - number
        - neighborhood
        - city
        - state
      properties:
        postalCode:
          type: string
          description: CEP
        street:
          type: string
          description: Logradouro
        number:
          type: string
          description: Número
        complement:
          type: string
          description: Complemento
        neighborhood:
          type: string
          description: Bairro
        city:
          type: string
          description: Cidade
        state:
          type: string
          description: Estado (UF)
    ProfessionalCouncil:
      type: object
      properties:
        type:
          type: string
          description: "Tipo do conselho profissional (ex: CRP, CRM)"
          enum:
            - CFEP
            - CRP
            - COREN
            - CRM
            - CREFITO
            - CREFONO
            - CRN
            - none
        registrationNumber:
          type: string
          description: Número de registro no conselho profissional
    ScreenerProfile:
      type: object
      required:
        - _id
        - cpf
        - firstName
        - surname
        - email
        - phone
        - address
        - professionalCouncil
        - jobTitle
        - degree
      properties:
        _id:
          type: string
        cpf:
          type: string
          description: CPF do Screener
        firstName:
          type: string
          description: Primeiro nome do Screener
        surname:
          type: string
          description: Sobrenome do Screener
        email:
          type: string
          format: email
          description: Endereço de e-mail do Screener
        phone:
          type: string
          description: Número de telefone do Screener
        address:
          $ref: '#/components/schemas/Address'
        professionalCouncil:
          $ref: '#/components/schemas/ProfessionalCouncil'
        jobTitle:
          type: string
          description: Cargo/profissão do Screener
        degree:
          type: string
          description: Formação acadêmica/grau do Screener
        darvCourseYear:
          type: integer
          description: Ano de conclusão do curso DARV (opcional)
    ScreenerModel:
      type: object
      required:
        - cpf
        - firstName
        - surname
        - email
        - password
        - phone
        - address
        - professionalCouncil
        - jobTitle
        - degree
      properties:
        _id:
          type: string
        cpf:
          type: string
          description: CPF do Screener
        firstName:
          type: string
          description: Primeiro nome do Screener
        surname:
          type: string
          description: Sobrenome do Screener
        email:
          type: string
          format: email
          description: Endereço de e-mail do Screener
        password:
          type: string
          description: Senha do Screener (hash)
        phone:
          type: string
          description: Número de telefone do Screener
        address:
          $ref: '#/components/schemas/Address'
        professionalCouncil:
          $ref: '#/components/schemas/ProfessionalCouncil'
        jobTitle:
          type: string
          description: Cargo/profissão do Screener
        degree:
          type: string
          description: Formação acadêmica/grau do Screener
        darvCourseYear:
          type: integer
          description: Ano de conclusão do curso DARV (opcional)
    ScreenerRegister:
      type: object
      required:
        - cpf
        - firstName
        - surname
        - email
        - password
        - phone
        - address
        - professionalCouncil
        - jobTitle
        - degree
      properties:
        cpf:
          type: string
          description: CPF do Screener
          example: '111.111.111-11'
        firstName:
          type: string
          description: Primeiro nome do Screener
          example: Maria
        surname:
          type: string
          description: Sobrenome do Screener
          example: Henriques Moreira Vale
        email:
          type: string
          format: email
          description: Endereço de e-mail do Screener
          example: maria.vale@holhos.com
        password:
          type: string
          description: Senha do Screener
          example: StrongPassword123
        phone:
          type: string
          description: Número de telefone do Screener
          example: '31988447613'
        address:
          $ref: '#/components/schemas/Address'
        professionalCouncil:
          $ref: '#/components/schemas/ProfessionalCouncil'
        jobTitle:
          type: string
          description: Cargo/profissão do Screener
          example: Psychologist
        degree:
          type: string
          description: Formação acadêmica/grau do Screener
          example: Psychology
        darvCourseYear:
          type: integer
          description: Ano de conclusão do curso DARV (opcional)
          example: 2019
    ScreenerLogin:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: Endereço de e-mail do Screener
          example: maria.vale@holhos.com
        password:
          type: string
          description: Senha do Screener
          example: StrongPassword123
    Token:
      type: object
      required:
        - access_token
        - token_type
      properties:
        access_token:
          type: string
        token_type:
          type: string
    ScreenerPasswordRecoveryRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          description: Endereço de e-mail do Screener para recuperação de senha
          example: maria.vale@holhos.com